+++ src/Http/Controllers/Lumen/Controller.php
@@ -4,8 +4,9 @@ namespace Tychovbh\Mvc\Http\Controllers\Lumen;
 
 use App\Http\Controllers\Rest;
 use Laravel\Lumen\Routing\Controller as BaseController;
+use App\Http\Controllers\Controller as IController;
 
-class Controller extends BaseController
+class Controller extends BaseController implements IController
 {
     use Rest;